Bugfix: Contexts weren't actually working.
How did we not catch this earlier?  There were two major bugs in the
implementation of Contexts.

First, Context.getStringVar never worked due to an incomplete definition
of the Compare funtion. (The consequence being that keys would be
sorted by string length, but for strings of equal length the wrong value
would be returned.)  Ugh.

Also, setStringVar (used in the Nuke context overrides) had no effect
due to a bug in the use of multimaps. Double ugh.