bugfix: symlinks in ocio profiles / lut paths are handled correctly.
Previously, there was incorrect handling of symlinks in lut paths:
(github issue #124)

If we had a pair of symlinks like:
   /tmp/ocio/nuke-default/config.ocio
   /tmp/ocio/nuke-default/luts/srgb.spi1d

Which point respectively to:
    /my/config/dir/config.ocio
    /my/luts/dir/srgb.spi1d
OCIO will complain saying that (after appending "luts" to the resolved config path):

The specified file reference 'srgb.spi1d' could not be located. The following attempts were made: /my/config/dir/luts/srgb.spi1d

This updates the behavior of OCIO, so the 'luts' path is specified relative to the initial path of the config, not the resolved symlink path.

The resource path code is also updated to support relative directories using the './' and '../' operators, which was previously not allowed.