core: added a new exception type, ExceptionMissingFile
This is also exposed in python as a custom exception type (along with the
normal, Exception).

This allows functions that want to remain resiliant to missing file errors,
(such as in LookTransform) to do so safely, where they will still fail
on more serious errors.

Nuke's OCIOLookTransform takes advantage of this to expose an 'ignoreErrors'
option.