cmake: soversion cleanup for both C++/python so(s)
Don't set SOVERSION on PyOpenColorIO plugin, on mac
Prevents "-compatibility_version only allowed with
-dynamiclib " error
From http://seclists.org/wireshark/2010/Sep/409 :

I'm not sure it's appropriate to set those versions for a plugin - they aren't shared libraries, so all the versioning stuff for shared libraries on various platforms still applies.  They shouldn't be installed with sonames - it should just be "asn1.so", not "asn1.so.1.5.0" with a symlink to "asn1.so" - and they shouldn't be built with -compatibility_version or -current_version on OS X, or with -soname on platforms using the GNU linker, or with +h on HP-UX with the HP linker, or with -h on Solaris with the Sun linker, and so on.

SOVERSION is always set in the top-level CMakeLists.txt, so no need to check `if(SOVERSION)` later (which doesn't always do as expected, if SOVERSION is set to zero)