core: first pass at internal optimization framework
The internal image-procesing ops now go through a simple optimization framework as part of Processor creation. In this first code pass, ops that are exact inverses of their neighbors cancel out and are removed.

But the most interesting consequence is that this allows for transforms which were not previously possible:

Say you have 2 colorspaces: A and B.  Both of these these colorspaces apply the same 3D lut, but B then also applies an additional matrix.

Previously, it would not have been possible to compute the processor to go directly from A to B because an inverse 3d lut is needed.
old op chain:   LUT3D (inv) -> LUT3D (forward) -> matrix

However, the new optimizer will detect that the same 3dlut is being applied in a row in both an inverse and forward direction, and they will cancel out! (This works even if the nesting is deep.)

new op chain: matrix

Subsequent additions to the optimizer will allow for matrix collapsing, etc.