core: added 'best' interpolation type
Also, specifying tetrahedral interpolation for lut1d throws an exception rather
than silently skipping lut application.

Addressed issue #217

The naming for 'BEST' was chosen (instead of alternatives such as HIGHEST or
SMOOTHEST) precisely as the ambiguity allows us a bit of freedom in which
interpolation method to use. Consider a 3d lut. Which is 'smoother'? Tetrahedral
or tri-linear? Both are smoother in certain combinations of input luts / image
content.  Thus, by use the intentially ambigious term 'best' we can punt on
these issues.

Of course, should eventually we add support for a cubic 1d interpolation type,
we will switch 'best' be CUBIC at that time.