looks syntax to allow for fallbacks
Example: "-cc,+di | +di | "

Note the introduction of the 'or' operator.

This example would use the '-cc,+di' look, if it exists without error.
If this look does not resolve (due to a missing file), then the "+di"
look will be used. If that doesnt succeed, then a 'no look' will be
used. (Note the trailing |). If one didnt want the 'no look' option to
be used as a fallback, the last | operator would be omitted.

We also expose a function, so given a processor, one ca
query which look(s) were actually used.

TODO: overall correctness testing, with regards to src,dst swapping for inverse
direction
TODO: test that intermediate colorspaces are the right value when look options
are used, and the first one errors.
todo: better accumulated error message for multiple look tries?
TODO: add an enum for 'nolook'?
TODO: in nuke, make OCIOLookApply allow arbitrary looks strings