nuke: config-dependent knob defaults are preserved
Previously, the knob value was only saved when set to a non-default value. For
values which are configuration dependent (such as colorspace names) this is not
ideal.

OCIOLogConvert direction is also stored, as we may wish to change the default
value in a future version and this will allow us.

OCIOLookTransform is updated in a different commit.