4ByteStruct configurations
Funny story! I built the AE plug-in with the default struct alignment,
not the 4 Byte version that AE plug-ins are supposed to use.
Interestingly, it does not seem to cause any problems, although
switching it now will cause problems with current projects.

Then there's the question of which version behaves well with projects
coming from the Mac. We assume the current one?

Anyway, 4ByteStruct configurations are there in case we ever decide to
use them.