After Effects update, Photoshop plug-in (#476)
* Stop accepting .ccc files while we don't have a way to get the ID

* Getting video space buffers for Premiere

* Check for empty paths

Seems like the smart thing to do, eh?

* Manually update version

* #define OLDYAML; the new YAML requires Boost so we'll avaoid it on Mac

* Manually update version

* Switch to new YAML

Wave of the future, dude.

* Add ext/.gitignore

* Add FileFormatCDL.cpp

* Add FileFormatCDL.cpp on win

* Add hierarchical menu for color spaces

* New menu for Windows

* Photoshop plug-in added

* Small changes to the AE plug-in for integration with PS

* 4ByteStruct configurations

Funny story! I built the AE plug-in with the default struct alignment,
not the 4 Byte version that AE plug-ins are supposed to use.
Interestingly, it does not seem to cause any problems, although
switching it now will cause problems with current projects.

Then there's the question of which version behaves well with projects
coming from the Mac. We assume the current one?

Anyway, 4ByteStruct configurations are there in case we ever decide to
use them.

* Photoshop plug-in building on Windows

* Mac fix

Found what appears to be a little bug on Mac while porting to Windows.

* Need to include OCIO if I want the version

* Fix output paths

* Fixed bug getting display transform

* Fix Windows crash

* extra line

* De-Tab