Add Academy/ASC Common LUT Format (CLF) read support (#36)
This branch adds full read support for the Academy/ASC Common LUT Format (CLF), Academy S-2014-006, along with new and improved processing ops.

New Range and ASC CDL ops have been added along with major enhancements to the Lut1D and Lut3D ops:  
-- The Range op provides easy level conversions and clamping.  
-- The ASC CDL op was added to allow implementation of various commonly used CDL formulas including one that is now faithful to the official ASC 1.2 spec (which OCIO's original implementation did not match).
-- Lut1D now supports the half-domain feature of CLF which allows a 1D LUT to be built using the same domain as a half-precision floating-point number.
-- Lut3D now supports inversion, meaning that all ops and color space conversions may now be applied in either direction.  
-- By default, the LUT inversion is using the fast approximation. Later a flag would select the fast or exact (but slow) LUT inversion algorithm.

This branch introduces an OpData layer containing the detailed operator machinery and unit tests and which allowed us to avoid making the core/*Op modules too large.   This code is in a new core/opdata directory.

The CPU renderers for the Lut1D, Lut3D, and matrix ops is now using new more optimized code.  This is now in a new core/cpu directory rather than in the core/*Op modules.  Support for integer bit-depths is still in progress and will be completed in a follow-on branch.  In the current branch, INTERP_NEAREST for Lut1D and Lut3D is temporarily handled as linear.

A core/clf directory was added containing the parsing code for the CLF format.  This makes use of the Expat XML parser which is faster and allows handling large files with a smaller memory footprint than TinyXML.

Many unit tests were added, including tests that validate the GPU renderer against the CPU renderer.

The ocioconvert and ociodisplay apps were updated:
-- ocioconvert now allows rendering with the GPU when compiled with the BUILD_WITH_GPU flag.
-- ociodisplay now allows use of the new GPU renderer.

The half module from ILMBase and the Expat library have been added under ext and the LICENSE file updated.