Improve the unit test capabilities and outputs (#564)
Improve unit test capabilities and outputs with new macros.


The pull request adds new macros and improves the unit test outputs.

The new macros stop the unit test if there is an error because it invalidates the following validations. Please refer the first unit test of src/core/MatrixOps.cpp to see the usage.

Note: After several changes in the src/oiio/src/include/unittest.h, the file is more and more specific to the OCIO project, and so it could be moved/renamed to src/core_tests or another existing framework could be used (with much more capabilities).