Core Infrastructure Initiative Best Practices status
Copied from here: https://github.com/coreinfrastructure/best-practices-badge/blob/master/doc/criteria.md

I added the colorful emoji annotations next to each item. Most are
marked with the 'unknown' question mark indicating that somebody more
knowledgeable needs to assess it, but I tried to mark the ones as done
(green check box) that were obvious to me.

The idea is that once we check this in to the developer docs, people can
submit PRs that modify the status of items that they know about, or as
they are fixed.