Adsk Contrib - const shared_ptr. (#628)
* const shared_ptr.
A const shared_ptr does not prevent to modify the pointed object. shared_ptr to const objects should be used instead.
This is adding const correctness. One drawback is that when a shared_ptr to data can not be used as a parameter to a function that expects a reference shared_ptr to const data. A shared_ptr to const data has to be constructed and passed to the function.

* On linux opData() functions are returning an object that needs to be assigned before being passed as a reference.

* Adjust Lut3D that got merged from master to use shared pointers to const objects