Adsk Contrib - Lut1D improvements (WIP) (#640)
* Lut1D improvements (WIP)

This PR contains upgrades to the Lut1D that speeds up inversion and sets the foundation for further improvements.  This PR is related to several upcoming PRs and is a Work in Progress.  As previously mentioned we are also planning on adding support for integer pixel buffers for the apply method.  This PR contains some code paths in Lut1DOpCPU.cpp that will not do anything until the bit-depth PR arrives and it will be substantially refactored at that time.  We are also adding support for CLF reading and until that arrives we cannot add some of the SynColor unit tests.

This PR keeps the original Lut1D struct from OCIO v1, which is used in the file format readers.  However in a future PR those will be switched over to using the new Lut1DOpData class and the original struct will be removed.  The code for determining if a LUT is a no-op/identity is changing somewhat as part of this and the new algorithm is in this PR but will not actually be used in the file readers until the upcoming PR.

Apologies for the incomplete aspects of this review.  We are trying to keep the PRs smaller but this implies the complete picture will not be clear until all the branches are in place.

* Update Lut1DOp.cpp

Replace tab by spaces

* Update Lut1DOpCPU.cpp

Update copyright

* Update src/OpenColorIO/ops/Lut1D/Lut1DOpData.cpp

Co-Authored-By: BernardLefebvre <37628108+BernardLefebvre@users.noreply.github.com>

* Update src/OpenColorIO/ops/Lut1D/Lut1DOpData.cpp

Co-Authored-By: BernardLefebvre <37628108+BernardLefebvre@users.noreply.github.com>

* Update src/OpenColorIO/ops/Lut3D/Lut3DOpData.cpp

Co-Authored-By: BernardLefebvre <37628108+BernardLefebvre@users.noreply.github.com>

* Remove Lut3D changes from the Lut1D PR.
Fix a compilation conversion warning.