Update CLAs per Sony legal department edits (#658)
The earlier one was copied straight from OpenVDB, which Sony Legal
decided was too closely tied to OpenVDB's use of the Mozilla Public
License.

This draft brings it more in line with the standard Linux Foundation
CLAs found here:
 https://github.com/AcademySoftwareFoundation/tac/tree/master/process

The new CCLA is exactly the LF one found above (with the proper name and
license substitutions). The new ICLA is very close to the CCLA, whereas
the LF template is more minimal; Sony Legal liked how the LF CCLA
spelled out some definitions more precisely and modeled after that.