Adsk Contrib - CLF parsing (#656)
This PR implements Issue #610.  It adds a new FileFormat to read Academy/ASC Common LUT Format (CLF) files.  This fully implements the S-2014-006 spec from the Academy.

Note: you will see references to CTF in the code.  The CLF support is based on the SynColor CTF reader, since CLF is a subset of CTF.  The Autodesk Color Transform Format (CTF) is a superset of CLF that adds support for all OCIO ops (e.g. including Exponent and Log) as well as some ops that are useful for accurate ACES implementations.  Because CTF requires adding some additional ops to OCIO, this PR does not enable CTF support yet.  That will come in a future PR.

This PR adds a lot of unit tests and the tests/data/unittestsfiles.tar.gz has many new example CLF files for parsing tests.  There are files that excercise all parts of the spec, including good files and files with syntax errors to test the exception handling.  We put a fair amount of work into trying to provide useful error messages when a file cannot be parsed, indicating the line number and the exact nature of the issue.

* Array function already declared in base class. Removes an override warning.

* Adjust exception messages to constantly say CTF/CLF. <Info> element is CTF only.

* Update CTFReaderHelper.cpp

Fix expression with side effects will be evaluated despite being used as an operand to 'typeid' [-Werror,-Wpotentially-evaluated-expression]