Restore C++03 compat when using old OIIO (#682)
We recently tried to change the OCIO code so that it would build
against both OIIO 2.0 as well as the older 1.x versions. In doing so,
we inadvertently broke OCIO for C++03.

This patch still requires C++11 when building against OIIO 2.x (it has
to), but is fine with C++03 if using older OIIO.
(cherry picked from commit 65a08dbedb771adc62b35b71e2a8b637776b8504)