Cherry picked fixes for proposed 1.1.x release (#678)
* Fix wrong display returned by getDefaultDisplay() (#638)

* Fix window-specific issue in GetEnv()

While the content of result was correct, its length was set
to a wrong size.

This is a patch by Patrick Hodoul.

* Fix wrong display returned by getDefaultDisplay()

Solved by ignoring environment variable if it's an empty string.
Ideally, need to test it to be NULL instead to support obscure
case when empty string is passed. Unfortunately, this is not
possible with the current API, and is unlikely scenario anyway.

The issue was caused by wrong logic around parsing display/views
override environment variable: if it is missing, it was converted
to an empty string. That empty string was considered an override
of one element display, with an empty name.

Additionally, make it so getDisplay() returns displays in order
of either active_displays configuration variable or environment
variable override. the issue there was caused by wrong order of
arguments to vector intersection.

Conflicts:
	src/OpenColorIO/Platform.cpp
	src/core/Config.cpp

This commit was cherry-picked from 718cc7f9f7. Changes were made
for compatibility with the older version of Platform.cpp; mainly
reverting "Getenv" to "getenv".

* Add Compatibility with OIIO 1.9+ API changes (#601)

This fix adds compatibility with OIIO 1.9+, minor ImageInput/ImageOutput API changes. At a certain point ImageInput::create switched from returning a raw pointer (that later needed to be deleted) to returning a std::unique_ptr.

* Correct ImageInput/ImageOutput swaps

* Restore C++03 compat when using old OIIO (#682)

We recently tried to change the OCIO code so that it would build
against both OIIO 2.0 as well as the older 1.x versions. In doing so,
we inadvertently broke OCIO for C++03.

This patch still requires C++11 when building against OIIO 2.x (it has
to), but is fine with C++03 if using older OIIO.
(cherry picked from commit 65a08dbedb771adc62b35b71e2a8b637776b8504)

* Add note about C++11 requirement when building with newer OIIO version

* Add error directive to apps when incompatable C++ std and OIIO lib are used