Update the CCLA with the extra words from the e-signature page.
One paragraph is additionally listed when you go to sign the actual
new CLA, which clarifies the role of the CLA manager. Adding it to the
copy of the CLA in our repo just so it exactly matches what people will
see and sign when they use the e-signature system.