Fix ociochecklut --gpu not working correctly for LUT1D
In OglApp::applyGPU, call to updateImage() update the currently bound GL_TEXTURE_2D
target, by doing so, the latter sampling of a 1DLUT that was allocated as
a GL_TEXTURE_2D (because of it's size), will sample the wrong texture.

Solution proposed is to call OglApp::setShader() again after image was updated.

Signed-off-by: Rémi Achard <remiachard@gmail.com>