Default file transform interpolation and transform types.
Changing the FileTransform default interpolation from unknown to default so that FileTransform can be used without changing the interpolation type.
Add interpolation validation to FileTransform validation.
Refactor how FileFormats are using the interpolation.
CTF/CLF uses the interpolation provided for LUTs that do not specifiy an interpolation within the file.
Other formats that can not specify interpolation in the file do remember the first interpolation used for a given file so that this interpolation can be reused if default interpolation is used.
Log a warning (instead of throw) if an interpolation can not be used.

Add a type to transform.

Signed-off-by: Bernard Lefebvre <bernard.lefebvre@autodesk.com>