Updates website with rebuilt version (#1330)
* Updates website with rebuilt version

- Rebuilt with `hugo -D --minify --themesDir ../..` as per the new docs
- Added .gitignore file for .DS_Store files in the `gh-pages` branch

This build notably still includes some un-used template blog posts which are also present in the live build as well as changes that I did not make like removal of old logos, an image from Spiderverse, and updates to application icons.  When I built the website on my local machine previously everything worked fine, assuming nothing has changed with the website config since it was last built (pre-build instructions) this should be good to go.  I recognize it is hard to review these types of changes.

Signed-off-by: Henry Wilkinson <henry@wilkinson.graphics>

* Updates sizzle reel

This will need to be updated in the main branch as well

Signed-off-by: Henry Wilkinson <henry@wilkinson.graphics>