Remove FindOpenImageIO.cmake
OCIO wants a minimum OIIO of 2.1, and coincidentally, that is the
release series in which OIIO started exporting OpenImageIOConfig.cmake
and friends, making the FindOpenImageIO module obsolete. This file is
not only unnecessary, but actually because it's out of date, it won't
work properly with upcoming OpenImageIO 2.3 (due to be released later
this year, or right now for anybody working from OIIO master branch).

Simply removing FindOpenImageIO.cmake will cause OCIO's build to use
the OIIO exported config.

I did notice that your OpenImageIO requirement was for 2.1.9, and there
is an unfortunate typo in that version's exported config. So I'm also
proposing making the minimum OIIO be 2.1.10, which was a patch released
in Jan 2020 -- a full year before OCIO v2 was released. OIIO continued
to issue 2.1 patches until 2.1.20 (Sept 2020), and since then, it's
actually OIIO 2.2.x that has recieved active support. I'm not saying
that you need to use 2.2 as a minimum, but I am trying to justify that
bumping your 2.1.9 minimum to 2.1.10 is not a radical move and will
avoid possible build troubles.

Signed-off-by: Larry Gritz <lg@larrygritz.com>