Bump CMake minimum version to 3.15, allow support for PIE linker flags
As per TSC discussion, we will bump minimum cmake version to 3.15,
this matches OpenImageIO and ASWF template project. Strictly speaking
CMake added this in 3.14, but matching the other ASWF users was seen
as a sensible option.

For correct support of PIE objects being linked correctly we need to
trigger cmake to check for pie support see:

https://cmake.org/cmake/help/latest/module/CheckPIESupported.html
https://cmake.org/cmake/help/latest/policy/CMP0083.html

Signed-off-by: Kevin Wheatley <kevin.wheatley@framestore.com>