Add support for environment variables in minimal configuration archives
If archiving a minimal config and environment variables are used, it must be possible for all the needed variables to be resolved and any files pointed to should also exist at the time of archiving.

The values of the environment will be written into the saved configuration in the environment section of the configuration file. Ths configuration stored thus may be different to the initial config being passed for archiving.

Needs tests adding and futher discussion in the TSC meetings.

Signed-off-by: Kevin Wheatley <kevin.wheatley@framestore.com>