Add cmake variable OCIO_INSTALL_OCIO
Defaults to ON.

This guards all of the install() commands, so of this variable is OFF,
no components of OCIO or its dependencies will be installed.

This is helpful when OCIO is not the "top level project", i.e. some
other project is building it as an unexposed static dependency and
doesn't want the OCIO components to be placed as artifacts in the
outer project's install.

Signed-off-by: Larry Gritz <lg@larrygritz.com>